Beta pre-release of python version
This is a pre-release of a semi-functional version of the compiler* written in Python. I decided to release these so I can say I made a programming language before I was 16 (now I only have a bit more than an hour, so i'm unlikely to debug the code). This is in no way shape or form something you would want to use. I believe I'll call it version 0.45 as it is not even half finished or usable. I do plan to finish the python compiler and also to make a version in C++ in the relatively near future (one or two months).

*the compiler is probably messed up as there is a little problem with the lexer and/or parser modules, as it cannot read the assignment (=) token properly and this doesn't allow the other parts of the code to be compiled.