Release of version 0.5
I created the environment module and updated the compiler to be up-to-date with the updates of other files. This will be the core of my program so I would count this to be half a version as there is much to work on but there is also much that has been done.

CS. A.